requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
sentence-transformers
chromadb
langchain
langchain-community
python-dotenv
